+++ packages/adp-tooling/src/source/applications.ts
@@ -1,8 +1,9 @@
-import { type AbapServiceProvider, AdaptationProjectType, type App } from '@sap-ux/axios-extension';
+import type { AbapServiceProvider, App } from '@sap-ux/axios-extension';
 import type { ToolsLogger } from '@sap-ux/logger';
 
 import { t } from '../i18n';
 import type { SourceApplication } from '../types';
+import { SupportedProject } from './systems';
 
 type UI5AppFilter = {
     fields: string;
@@ -10,7 +11,7 @@ type UI5AppFilter = {
     readonly ['sap.app/type']: 'application';
 } & Record<string, string>;
 
-const ONPREM_APP_FIELDS = [
+const ONPREM_APP_FIELDS_LIST = [
     'sap.app/id',
     'sap.app/ach',
     'sap.fiori/registrationIds',
@@ -20,17 +21,18 @@ const ONPREM_APP_FIELDS = [
     'repoName'
 ];
 
-const CLOUD_APP_FIELDS = [...ONPREM_APP_FIELDS, 'sap.fiori/cloudDevAdaptationStatus'];
+const ONPREM_APP_FIELDS_STR = ONPREM_APP_FIELDS_LIST.join(',');
+const CLOUD_APP_FIELDS_STR = [...ONPREM_APP_FIELDS_LIST, 'sap.fiori/cloudDevAdaptationStatus'].join(',');
 
 const APPS_WITH_DESCR_FILTER: UI5AppFilter = {
-    fields: ONPREM_APP_FIELDS.join(','),
+    fields: ONPREM_APP_FIELDS_STR,
     'sap.ui/technology': 'UI5',
     'sap.app/type': 'application',
     'fileType': 'appdescr'
 };
 
 const APPS_WITH_VARIANT_DESCR_FILTER: UI5AppFilter = {
-    fields: ONPREM_APP_FIELDS.join(','),
+    fields: ONPREM_APP_FIELDS_STR,
     'sap.ui/technology': 'UI5',
     'sap.app/type': 'application',
     'fileType': 'appdescr_variant',
@@ -38,7 +40,7 @@ const APPS_WITH_VARIANT_DESCR_FILTER: UI5AppFilter = {
 };
 
 const CLOUD_ONLY_APPS_FILTER: UI5AppFilter = {
-    fields: CLOUD_APP_FIELDS.join(','),
+    fields: CLOUD_APP_FIELDS_STR,
     'sap.app/type': 'application',
     'sap.fiori/cloudDevAdaptationStatus': 'released'
 };
@@ -114,19 +116,19 @@ export async function isAppSupported(provider: AbapServiceProvider, id: string, 
  *
  * @param {AbapServiceProvider} provider - The ABAP service provider used to retrieve application data.
  * @param {boolean} isCustomerBase - Flag indicating whether the system is customer-based. Affects application selection.
- * @param {AdaptationProjectType | undefined} projectType - The project type.
+ * @param {SupportedProject | undefined} supportedProject - The supported ADP projects by the system.
  * @returns {Promise<SourceApplication[]>} If the project type is cloudReady resolves with
  * applications for which the {@link SourceApplication.cloudDevAdaptationStatus} is `released`.
  * If the project type is onPremise we display all applications, plus applications with variant descriptor
- * in case the {@link isCustomerBase} flag is set to true. In case the {@link projectType} is NOT set
+ * in case the {@link isCustomerBase} flag is set to true. In case the {@link supportedProject} is NOT set
  * we return an empty array.
  */
 export async function loadApps(
     provider: AbapServiceProvider,
     isCustomerBase: boolean,
-    projectType?: AdaptationProjectType
+    supportedProject?: SupportedProject
 ): Promise<SourceApplication[]> {
-    if (!projectType) {
+    if (!supportedProject) {
         return [];
     }
 
@@ -134,18 +136,39 @@ export async function loadApps(
         const appIndexService = provider.getAppIndex();
 
         const appIndex = (
-            await Promise.all([
-                appIndexService.search(
-                    projectType === AdaptationProjectType.CLOUD_READY ? CLOUD_ONLY_APPS_FILTER : APPS_WITH_DESCR_FILTER
-                ),
-                projectType === AdaptationProjectType.ON_PREMISE && isCustomerBase
-                    ? appIndexService.search(APPS_WITH_VARIANT_DESCR_FILTER)
-                    : Promise.resolve([])
-            ])
+            await Promise.all(
+                getAppFilters(isCustomerBase, supportedProject).map((filter) => appIndexService.search(filter))
+            )
         ).flat();
 
         return appIndex.map(toSourceApplication).sort(compareByTitleOrId);
     } catch (error) {
         throw new Error(`Could not load applications: ${error.message}`);
     }
+}
+
+/**
+ * Helper method used to create the application filters required by the app index service.
+ *
+ * @param {boolean} isCustomerBase Indicates whether the system is customer-based or not.
+ * @param {SupportedProject} supportedProject The supported ADP projects by the system.
+ * @returns {UI5AppFilter[]} The array of filters dermined by the supported ADP projects from the system.
+ */
+function getAppFilters(isCustomerBase: boolean, supportedProject: SupportedProject): UI5AppFilter[] {
+    if (supportedProject === SupportedProject.CLOUD_READY) {
+        return [CLOUD_ONLY_APPS_FILTER];
+    }
+
+    const displayAllAppsFilters = [APPS_WITH_DESCR_FILTER, ...(isCustomerBase ? [APPS_WITH_VARIANT_DESCR_FILTER] : [])];
+
+    if (supportedProject === SupportedProject.CLOUD_READY_AND_ON_PREM) {
+        // In case of a mixed system we want to also include the cloudDevAdaptationStatus property
+        // for each application result in the list. The property is available as a column for all apps
+        // due to the nature of the system - mixed, so we are safe no error 400 will occur like
+        // with older onPremise only systems for which the column is missing. For non cloud app the
+        // property cloudDevAdaptationStatus has value an empty string for a cloud app - 'released'.
+        return displayAllAppsFilters.map((filter) => ({ ...filter, fields: CLOUD_APP_FIELDS_STR }));
+    }
+
+    return displayAllAppsFilters;
 }