+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/fe-fpm-cli
 
+## 0.1.130
+
+### Patch Changes
+
+- @sap-ux/fe-fpm-writer@0.43.16
+
 ## 0.1.129
 
 ### Patch Changes