+++ packages/project-input-validator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/project-input-validator
 
+## 0.6.72
+
+### Patch Changes
+
+- Updated dependencies [b66e827]
+    - @sap-ux/project-access@1.35.17
+
 ## 0.6.71
 
 ### Patch Changes