+++ packages/project-integrity/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/project-integrity
 
+## 0.2.64
+
+### Patch Changes
+
+- Updated dependencies [b66e827]
+    - @sap-ux/project-access@1.35.17
+
 ## 0.2.63
 
 ### Patch Changes