+++ packages/ui5-test-writer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/ui5-test-writer
 
+## 0.7.95
+
+### Patch Changes
+
+- aa2baf3: fix action button state test based on line item selection
+
+## 0.7.94
+
+### Patch Changes
+
+- Updated dependencies [b66e827]
+    - @sap-ux/project-access@1.35.17
+    - @sap-ux/ui5-application-writer@1.8.2
+
 ## 0.7.93
 
 ### Patch Changes